Use gitlab ci, reduce testload on travis

+++ tools/install_hooks.sh
@@ -8,8 +8,10 @@ fi
 
 cd $SCRIPTPATH/..
 
-rm -f .git/hooks/pre-commit
-ln -s $SCRIPTPATH/../tools/pre-commit-hook .git/hooks/pre-commit
+if [ -d .git/hooks ]; then
+    rm -f .git/hooks/pre-commit
+    ln -s $SCRIPTPATH/../tools/pre-commit-hook .git/hooks/pre-commit
 
-rm -f .git/hooks/pre-push
-ln -s $SCRIPTPATH/../tools/pre-push-hook .git/hooks/pre-push
+    rm -f .git/hooks/pre-push
+    ln -s $SCRIPTPATH/../tools/pre-push-hook .git/hooks/pre-push
+fi
